Created NPC Spawning
I created a spawner where you can set the amount of NPCs to spawn and the range the can spawn. I also made a function that picks one of the NPCs and assigns a task so you can talk to them. Right now there are no tasks but I am in the process of creating a class for it.