WIP Able to talk to each NPC
I have made it so now you can talk to each npc as long as they have dialogue and a scene. I also made it so the player so placed next to them and the camera zooms in on them. Right now it has an error when coming back from a game.